pyproject.toml
[tool.poetry]
name = "shranbot"
version = "1.0.0"
description = "Discord bot for Shran development discord server"
authors = ["Matt Williams <matt@dmgblockchain.com>"]

[tool.poetry.dependencies]
python = "^3.8"
colorlog = "^4.7.2"
colorama = "^0.4.4"
python-dotenv = "^0.17.0"
"discord.py" = "^1.7.3"
result = "^0.6.0"
yfinance = "^0.1.66"
emoji = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.812"
jedi = "^0.18.0"
isort = "^5.7.0"
pylint = "^2.7.2"
autopep8 = "^1.5.7"

[tool.poetry.scripts]
drive = "shranbot.main:run"
tests = "scripts.project_helper:run_tests"
stubs = "scripts.project_helper:stub_gen"
analyze = "scripts.project_helper:run_analyzer"

[build-system]
requires = ["poetry-core>=1.0.0"]

build-backend = "poetry.core.masonry.api"
